Homework-4
Есть несколько вопросов.
1) Как ты именуешь переменные, есть ли какой то подход что бы не путаться.
2) Я нашел баг но не знаю пока как его устранить и не понимаю почему он так работает. В общем я перенес функционал которые перебрасывают на страницу game и home в меню из третего задания. И если нажать на странице Home на кнопку которая отвечает за переход на эту страницу эта функция перестает работать.
Спасибо, за урок.
