+++ src/components/Footer/index.js
@@ -1,9 +1,9 @@
-import style from './style.module.css'
+import s from './style.module.css'
 
 const Footer = () => {
     return (
         <footer>
-            <div className={style.wrapper}>
+            <div className={s.wrapper}>
                 <h3>THANKS FOR VISITING</h3>
                 <p>© 2021 #ReactMarathon.</p>
             </div>
