+++ src/components/Header/index.js
@@ -1,13 +1,15 @@
-import style from './style.module.css'
+import s from './style.module.css'
 
-const Header = (props) => {
+const Header = ({title, desc}) => {
+
     return (
-        <header className={style.root}>
-            <div className={style.forest}></div>
-            <div className={style.container}>
-                <h1>{props.title}</h1>
-                <p>{props.desc}</p>
+        <header className={s.root}>
+            <div className={s.forest}></div>
+            <div className={s.container}>
+                <h1>{title}</h1>
+                <p>{desc}</p>
             </div>
+
         </header>
     )
 }
