+++ src/components/PokemonCard/index.js
@@ -1,6 +1,9 @@
-import s from './style.module.css'
-import image from './assets/card-back-side.jpg'
+import cn from 'classnames'
 import { useState } from 'react'
+
+import image from './assets/card-back-side.jpg'
+
+import s from './style.module.css'
 
 
 
@@ -9,20 +12,20 @@ const PokemonCard = ({name, id, type, img, values}) => {
     const [isActive, setActive] = useState(false)
 
     const onCard = () => {
-        setActive(true)
+        setActive(!isActive)
     }
 
     return (
         <div className={s.root}>
-            <div onClick={onCard} className={`${s.pokemonCard} ${isActive ? s.active : ''}`}>
+            <div onClick={onCard} className={cn(s.pokemonCard, {[s.active]: isActive})}>
                 <div className={s.cardFront}>
-                    <div className={`${s.wrap} ${s.front}`}>
-                        <div className={`${s.pokemon} ${s[type]}`}>
+                    <div className={cn(s.wrap, s.front)}>
+                        <div className={cn(s.pokemon, s[type])}>
                             <div className={s.values}>
-                                <div className={`${s.count} ${s.top}`}> {values.top} </div>
-                                <div className={`${s.count} ${s.right}`}> {values.right} </div>
-                                <div className={`${s.count} ${s.bottom}`}> {values.bottom} </div>
-                                <div className={`${s.count} ${s.left}`}> {values.left} </div>
+                                <div className={cn(s.count, s.top)}> {values.top} </div>
+                                <div className={cn(s.count, s.right)}> {values.right} </div>
+                                <div className={cn(s.count, s.bottom)}> {values.bottom} </div>
+                                <div className={cn(s.count, s.left)}> {values.left} </div>
                             </div>
                             <div className={s.imgContainer}>
                                 <img src={img} alt={name}/>
@@ -30,14 +33,14 @@ const PokemonCard = ({name, id, type, img, values}) => {
                             <div className={s.info}>
                                 <span className={s.number}>#{id}</span>
                                 <h3 className={s.name}> {name} </h3>
-                                <small className={s.type}>Type:<span> {type} </span></small>
+                                <small>Type:<span> {type} </span></small>
                             </div>
                         </div>
                     </div>
                 </div>
 
                 <div className={s.cardBack}>
-                    <div className={`${s.wrap} ${s.back}`} >
+                    <div className={cn(s.wrap, s.back)} >
                         <img src={image} alt="Сard Backed"/>
                     </div>
                 </div>
