+++ src/routes/Game/style.module.css
@@ -1,0 +1,3 @@
+.game {
+    overflow: hidden;
+}
